backend/requirements.txt
fastapi
uvicorn
pymongo
requests
boto3
openai
google-generativeai
scikit-learn
python-dotenv
pydantic==2.11.2